EC# lexer: remove HashId/NormalId distinction
A rarely-used feature of the lexer is that if the first character is
punctuation then the character set of the identifier is expanded
beyond [a-zA-Z0-9_'#] to include [~!%^*\\-+=|<>/?:.$].

I'd forgotten about this, so I considered removing it, but it seems
vaguely beneficial to be able to write things like `@'==` in
documentation and have it be legal syntax in both EC# and LES2.

But being able to abbreviate @`#<foo>` as @#<foo> has almost no
value, as it's very rare for a symbol to start with `#` and also contain
punctuation, so I'm removing HashId rule that allowed this.

Also:
- Tweaks
- Increased version to 2.7.0.1 to avoid confusing myself about which
  unreleased version of LeMP I'm using.