Simplify `IListSource.Slice` return value to `IListSource`
Experience has shown that Ranges don't work as well in C# as they do
in D, and changing the return value to IListSource required no changes
to any of the consumers of the lists in this codebase.

Later I might remove IRange and some of its methods from the Slice
classes, but this commit doesn't do that.